requirements.txt
numpy
scipy
pandas
opencv-python
Pillow
torchvision
torchvision
facenet-pytorch
faiss-cpu
scikit-learn
tqdm